Move utility functions out so they can be reused between tests
