+++ tests/test_pyrevolve.py
@@ -1,46 +1,7 @@
-from pyrevolve import Operator, Checkpoint, Revolver
+from utils import SimpleOperator, SimpleCheckpoint
+from pyrevolve import Revolver
 
-import numpy as np
 import pytest
-
-
-def np_ref_address(ptr):
-    return ptr.__array_interface__['data'][0]
-
-
-class SimpleCheckpoint(Checkpoint):
-    def __init__(self):
-        self.save_counter = 0
-        self.load_counter = 0
-        self.save_pointers = set()
-        self.load_pointers = set()
-
-    def save(self, ptr):
-        self.save_counter += 1
-        self.save_pointers.add(np_ref_address(ptr))
-
-    def load(self, ptr):
-        self.load_counter += 1
-        self.load_pointers.add(np_ref_address(ptr))
-
-    @property
-    def dtype(self):
-        return np.float32
-
-    @property
-    def size(self):
-        return 10
-
-
-class SimpleOperator(Operator):
-    def __init__(self):
-        self.counter = 0
-
-    def apply(self, *args, **kwargs):
-        t_start = kwargs['t_start']
-        t_end = kwargs['t_end']
-        self.counter += abs(t_end - t_start)
-
 
 @pytest.mark.parametrize("nt, ncp", [(10, 2), (10, 4), (10, 6), (10, 8),
                                      (10, 9), (10, 10), (10, 11), (10, 12)])
